pyproject.toml
[project]
name = "opencode-mastery"
version = "1.2.0"
description = "OpenCode knowledge base with lazy-loaded docs, fuzzy search and intelligent memory"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[dependency-groups]
dev = [
    "python-frontmatter>=1.1.0",
]
